+++ src/containers/PageConstructor/__stories__/components/CustomCard/index.ts
@@ -1,0 +1,1 @@
+export {CustomCard, CustomCardProps} from './CustomCard';