+++ src/containers/PageConstructor/__stories__/components/CustomHeader/index.ts
@@ -1,0 +1,1 @@
+export {CustomHeader} from './CustomHeader';